refactor: colocate achievement unlock service tests
